вынес константы в отдельный модуль
